+++ dao/announcement.js
@@ -1,5 +1,5 @@
 /**
-  * afterloe - cynomy_portal_server/dao/notice.js
+  * afterloe - cynomy_portal_server/dao/announcement.js
   *
   * Copyright(c) afterloe.
   * MIT Licensed
@@ -7,7 +7,7 @@
   * Authors:
   *   afterloe <lm6289511@gmail.com> (https://github.com/afterloe)
   * Date:
-  *   2017-3-9 15:32:43
+  *   2017-3-10 14:52:46
   */
 "use strict";
 
@@ -19,6 +19,6 @@ const classMethod = {
 
 Object.assign(commonsLib, classMethod);
 
-const className = "notice";
+const className = "announcement";
 
 module.exports = _ => _.definition({classMethod: commonsLib, className});